requirements_dev.txt
pip==19.2.3
bump2version==0.5.11
wheel==0.33.6
watchdog==0.9.0
tox==3.14.0
coverage==4.5.4
Sphinx==7.2.6
twine==5.0.0
ruff==0.3.5
google-genai == 1.24.0
playwright == 1.53.0
beautifulsoup4 == 4.13.4
pytest==8.4.1
